[SYCL][ABI-Break] Merge DeviceBinaryImage and RTDeviceBinaryImage
To avoid future issues with ABI for DeviceBinaryImage, this commit makes
the following chages:
* Moves RTDeviceBinaryImage to a source header.
* Merges DeviceBinaryImage into RTDeviceBinaryImage.
* Promotes the common property ranges from DeviceBinaryImage into new
  members.
