+++ sycl/source/detail/program_manager/program_manager.cpp
@@ -383,7 +383,7 @@ static void appendLinkOptionsFromImage(std::string &LinkOpts,
 static bool getUint32PropAsBool(const RTDeviceBinaryImage &Img,
                                 const char *PropName) {
   pi_device_binary_property Prop = Img.getProperty(PropName);
-  return Prop && (pi::DeviceBinaryProperty(Prop).asUint32() != 0);
+  return Prop && (DeviceBinaryProperty(Prop).asUint32() != 0);
 }
 
 static void appendCompileOptionsFromImage(std::string &CompileOpts,
@@ -1084,7 +1084,7 @@ ProgramManager::build(ProgramPtr Program, const ContextImplPtr Context,
 }
 
 static ProgramManager::KernelArgMask
-createKernelArgMask(const pi::ByteArray &Bytes) {
+createKernelArgMask(const ByteArray &Bytes) {
   const int NBytesForSize = 8;
   const int NBitsInElement = 8;
   std::uint64_t SizeInBits = 0;
@@ -1102,7 +1102,7 @@ createKernelArgMask(const pi::ByteArray &Bytes) {
 
 void ProgramManager::cacheKernelUsesAssertInfo(OSModuleHandle M,
                                                RTDeviceBinaryImage &Img) {
-  const pi::DeviceBinaryImage::PropertyRange &AssertUsedRange =
+  const RTDeviceBinaryImage::PropertyRange &AssertUsedRange =
       Img.getAssertUsed();
   if (AssertUsedRange.isAvailable())
     for (const auto &Prop : AssertUsedRange) {
@@ -1129,14 +1129,14 @@ void ProgramManager::addImages(pi_device_binaries DeviceBinary) {
     auto Img = make_unique_ptr<RTDeviceBinaryImage>(RawImg, M);
 
     // Fill the kernel argument mask map
-    const pi::DeviceBinaryImage::PropertyRange &KPOIRange =
+    const RTDeviceBinaryImage::PropertyRange &KPOIRange =
         Img->getKernelParamOptInfo();
     if (KPOIRange.isAvailable()) {
       KernelNameToArgMaskMap &ArgMaskMap =
           m_EliminatedKernelArgMasks[Img.get()];
       for (const auto &Info : KPOIRange)
         ArgMaskMap[Info->Name] =
-            createKernelArgMask(pi::DeviceBinaryProperty(Info).asByteArray());
+            createKernelArgMask(DeviceBinaryProperty(Info).asByteArray());
     }
 
     // Fill maps for kernel bundles
@@ -1226,8 +1226,8 @@ void ProgramManager::addImages(pi_device_binaries DeviceBinary) {
 
         auto DeviceGlobals = Img->getDeviceGlobals();
         for (const pi_device_binary_property &DeviceGlobal : DeviceGlobals) {
-          pi::ByteArray DeviceGlobalInfo =
-              pi::DeviceBinaryProperty(DeviceGlobal).asByteArray();
+          ByteArray DeviceGlobalInfo =
+              DeviceBinaryProperty(DeviceGlobal).asByteArray();
 
           // The supplied device_global info property is expected to contain:
           // * 8 bytes - Size of the property.
@@ -1391,10 +1391,10 @@ void ProgramManager::flushSpecConstants(const program_impl &Prg,
 // mask, sycl runtime won't know which fallback device libraries are needed. In
 // such case, the safest way is to load all fallback device libraries.
 uint32_t ProgramManager::getDeviceLibReqMask(const RTDeviceBinaryImage &Img) {
-  const pi::DeviceBinaryImage::PropertyRange &DLMRange =
+  const RTDeviceBinaryImage::PropertyRange &DLMRange =
       Img.getDeviceLibReqMask();
   if (DLMRange.isAvailable())
-    return pi::DeviceBinaryProperty(*(DLMRange.begin())).asUint32();
+    return DeviceBinaryProperty(*(DLMRange.begin())).asUint32();
   else
     return 0xFFFFFFFF;
 }
