+++ sycl/test/abi/vtable.cpp
@@ -23,16 +23,6 @@ void foo(sycl::detail::HostKernelBase &HKB) {
 // CHECK-NEXT:   4 | sycl::detail::HostKernelBase::~HostKernelBase() [complete]
 // CHECK-NEXT:   5 | sycl::detail::HostKernelBase::~HostKernelBase() [deleting]
 
-void foo(sycl::detail::pi::DeviceBinaryImage &Img) { Img.print(); }
-// CHECK:    Vtable for 'sycl::detail::pi::DeviceBinaryImage' (6 entries).
-// CHECK-NEXT:   0 | offset_to_top (0)
-// CHECK-NEXT:   1 | sycl::detail::pi::DeviceBinaryImage RTTI
-// CHECK-NEXT:       -- (sycl::detail::pi::DeviceBinaryImage, 0) vtable address --
-// CHECK-NEXT:   2 | void sycl::detail::pi::DeviceBinaryImage::print() const
-// CHECK-NEXT:   3 | void sycl::detail::pi::DeviceBinaryImage::dump(std::ostream &) const
-// CHECK-NEXT:   4 | sycl::detail::pi::DeviceBinaryImage::~DeviceBinaryImage() [complete]
-// CHECK-NEXT:   5 | sycl::detail::pi::DeviceBinaryImage::~DeviceBinaryImage() [deleting]
-
 void foo(sycl::detail::CG *CG) { delete CG; }
 // CHECK:    Vtable for 'sycl::detail::CG' (4 entries).
 // CHECK-NEXT:   0 | offset_to_top (0)
